blog/requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-binary-database-files==1.0.15
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-import-export==2.6.1
django-js-asset==1.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
